third_party/WebKit/LayoutTests/webaudio/pannernode-basic-expected.txt

CONSOLE WARNING: The provided value '1' is not a valid enum value of type PanningModelType.
CONSOLE WARNING: The provided value '1' is not a valid enum value of type DistanceModelType.
Basic tests for PannerNode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PannerNode has one input.
PASS PannerNode has one output.
PASS refDistance default value is 1.
PASS refDistance value is set to 270.52.
PASS maxDistance default value is 10000.
PASS maxDistance value is set to 100.55.
PASS rolloffFactor default value is 1.
PASS rolloffFactor value is set to 0.83.
PASS coneInnerAngle default value is 360.
PASS coneInnerAngle value is set to 240.45.
PASS coneOuterAngle default value is 360.
PASS coneOuterAngle value is set to 166.66.
PASS coneOuterGain default value is 0.
PASS coneOuterGain value is set to 0.35.
PASS PannerNode defaults to 'equalpower' panningModel.
PASS PannerNode defaults to 'inverse' distanceModel.
PASS panningModel: 'equalpower' is settable.
PASS panningModel: 'HRTF' is settable.
PASS panningModel correctly not set to 1.
PASS distanceModel: 'linear' is settable.
PASS distanceModel: 'inverse' is settable.
PASS distanceModel: 'exponential' is settable.
PASS distanceModel correctly not set to 1.
PASS successfullyParsed is true

TEST COMPLETE
